Conflicts:
	src/client/zeusclient/control/controlobject/out_text.cpp

+++ src/client/zeusclient/control/controlobject/out_text.cpp
@@ -5,14 +5,13 @@ int PrintText::m_Pos = 0;
 
 PrintTextFont* PrintText::m_Font = NULL;
 
-bool PrintText::Print(int x, int y,
+bool PrintText::Print(float x, float y,
         TextRect rect, char* strText, ...)
 {
     //无资源池 暂时这样
     int nBufferLen = strlen(strText) + 80;
     char* szBuffer = new char[nBufferLen];
     memset(szBuffer, 0, nBufferLen);
-
     m_Font = new PrintTextFont("res\\font1.fnt");
 
     va_list argptr;  //参数列表
